Cleaned code in PointMapperImpl, refactored classes to corresponding packages
